Make maxScore double instead of int

+++ src/main/java/ch/uzh/ifi/access/course/model/ExerciseConfig.java
@@ -15,7 +15,7 @@ public class ExerciseConfig implements Serializable {
     protected ExerciseType type;
     protected String language;
     protected Boolean isGraded;
-    protected int maxScore;
+    protected double maxScore;
     protected int maxSubmits;
     protected String gradingSetup;
